requirements.txt
Flask==2.0.2
gunicorn
azure-storage-blob==12.16.0
